expr exzero: take value type as a 'long int'
Depending on the platform, Flex can generate a `long int` for this field. This
results in 3 -Wconversion warnings during compilation due to `exzero` calls. We
can suppress these by using the widest type Flex might use in `exzero`.

+++ lib/expr/expr.h
@@ -276,7 +276,7 @@ extern int		extoken_fn(Expr_t*);
 extern char*		exstring(Expr_t *, char *);
 extern void*		exstralloc(Expr_t *, size_t);
 extern char*		extype(int);
-extern Extype_t		exzero(int);
+extern Extype_t exzero(long int);
 extern char*	exopname(int);
 extern void		exinit(void);
 extern char*	extypename(Expr_t * p, int);
